Sonar Commit
Commit for resolving Sonar Cloud issues
