+++ src/components/general/mediaTypes/ImageViewer.tsx
@@ -1,7 +1,6 @@
 /* Import Dependencies */
 import { useEffect, useState, useRef } from 'react';
 import KeycloakService from 'keycloak/Keycloak';
-import { Formik, Form, Field } from 'formik';
 import { isEmpty } from 'lodash';
 import {
     AnnotoriousOpenSeadragonAnnotator,
@@ -12,8 +11,6 @@ import {
     PointerSelectAction,
     useAnnotator
 } from '@annotorious/react';
-import classNames from 'classnames';
-import { Row, Col } from 'react-bootstrap';
 
 /* Import Store */
 import { useAppSelector, useAppDispatch } from "app/hooks";
@@ -23,17 +20,13 @@ import { getDigitalMedia, getDigitalMediaAnnotations } from "redux/digitalMedia/
 /* Import Types */
 import { Annotation, ImageAnnotationTemplate, Dict } from 'global/Types';
 
-/* Import Styles */
-import styles from 'components/general/mediaTypes/mediaTypes.module.scss';
-
-/* Import Icons */
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPencil, faTrashCan } from '@fortawesome/free-solid-svg-icons';
-
 /* Import API */
 import InsertAnnotation from 'api/annotate/InsertAnnotation';
 import PatchAnnotation from 'api/annotate/PatchAnnotation';
 import DeleteAnnotation from 'api/annotate/DeleteAnnotation';
+
+/* Import Components */
+import ImagePopup from './ImagePopup';
 
 
 /* Props Typing */
@@ -185,7 +178,7 @@ const ImageViewer = (props: Props) => {
             /* For each annotation, calculate the W3C pixels relative to the TDWG AC position */
             digitalMediaAnnotations.visual.forEach((imageAnnotation: Annotation) => {
                 /* Check if the annotation is a visual one */
-                if (imageAnnotation.target.selector && imageAnnotation.target.selector.hasROI) {
+                if (imageAnnotation.target.selector?.hasROI) {
                     const ROI = imageAnnotation.target.selector.hasROI;
 
                     const x = ROI["ac:xFrac"] * imgWidth;
@@ -275,17 +268,6 @@ const ImageViewer = (props: Props) => {
         }
     }
 
-    /* Function for validating Annotation value input */
-    const ValidateAnnotation = (values: Dict) => {
-        const errors: Dict = {};
-
-        if (!values.annotationValue) {
-            errors.annotationValue = "Empty annotations are not accepted";
-        }
-
-        return errors;
-    }
-
     /* Function for removing an Annotation */
     const RemoveAnnotation = () => {
         /* Ask for user confirmation */
@@ -298,15 +280,6 @@ const ImageViewer = (props: Props) => {
             });
         }
     }
-
-    /* ClassNames */
-    const classEditBlock = classNames({
-        'd-none': !editAnnotation
-    });
-
-    const classInfoBlock = classNames({
-        'd-none': editAnnotation
-    });
 
     return (
         <div className="w-100 h-100">
@@ -322,118 +295,17 @@ const ImageViewer = (props: Props) => {
 
                 <OpenSeadragonPopup
                     popup={() => (
-                        <div className={styles.annotoriousPopUp}>
-                            <Row>
-                                <Col className="bgc-white px-3 py-2">
-                                    {/* Annotation values */}
-                                    <Row className={`${styles.annotoriousPopUpSection} pb-2`}>
-                                        <Col>
-                                            <div className={classEditBlock}>
-                                                <Formik
-                                                    initialValues={{
-                                                        annotationValue: (editAnnotation && editAnnotation.body.length) ?
-                                                            editAnnotation.body[0].value[0] : '',
-                                                    }}
-                                                    enableReinitialize={true}
-                                                    validate={ValidateAnnotation}
-                                                    onSubmit={async (form) => {
-                                                        await new Promise((resolve) => setTimeout(resolve, 500));
+                        <ImagePopup selectedAnnotation={selectedAnnotation}
+                            editAnnotation={editAnnotation}
+                            tooltipFieldRef={tooltipFieldRef}
+                            anno={anno}
 
-                                                        /* Submit new Annotation */
-                                                        if (!selectedAnnotation.body.length) {
-                                                            /* Patch Annotation */
-                                                            SubmitAnnotation([form.annotationValue], 'insert');
-                                                        } else {
-                                                            /* Insert Annotation */
-                                                            SubmitAnnotation([form.annotationValue], 'patch')
-                                                        }
-                                                    }}
-                                                >
-                                                    {({ errors }) => (
-                                                        <Form>
-                                                            <Row>
-                                                                <Col>
-                                                                    <Field name="annotationValue"
-                                                                        type="text"
-                                                                        innerRef={tooltipFieldRef}
-                                                                        className="w-100 formField fs-4 py-1"
-                                                                    />
-
-                                                                    {errors.annotationValue &&
-                                                                        <p className="fs-5 c-denied mt-2">
-                                                                            {errors.annotationValue as string}
-                                                                        </p>
-                                                                    }
-                                                                </Col>
-                                                                <Col className="col-md-auto ps-0">
-                                                                    <button type="submit"
-                                                                        className="secondaryButton fs-4 px-3"
-                                                                    >
-                                                                        Save
-                                                                    </button>
-                                                                </Col>
-                                                            </Row>
-                                                        </Form>
-                                                    )}
-                                                </Formik>
-                                            </div>
-
-                                            <div className={classInfoBlock}>
-                                                <Row>
-                                                    <Col>
-                                                        <p className="fs-4">
-                                                            {(selectedAnnotation && selectedAnnotation.body.length) ?
-                                                                selectedAnnotation.body[0].value[0] : ''
-                                                            }
-                                                        </p>
-                                                    </Col>
-                                                    {(KeycloakService.IsLoggedIn() && selectedAnnotation && selectedAnnotation.body.length
-                                                        && KeycloakService.GetSubject() === selectedAnnotation.body[0].creator.id) &&
-                                                        <>
-                                                            <Col className="col-md-auto">
-                                                                <FontAwesomeIcon icon={faPencil}
-                                                                    className="c-pointer"
-                                                                    onClick={() => setEditAnnotation(selectedAnnotation)}
-                                                                />
-                                                            </Col>
-                                                            <Col className="col-md-auto ps-0">
-                                                                <FontAwesomeIcon icon={faTrashCan}
-                                                                    className="c-pointer"
-                                                                    onClick={() => RemoveAnnotation()}
-                                                                />
-                                                            </Col>
-                                                        </>
-                                                    }
-                                                </Row>
-                                            </div>
-                                        </Col>
-                                    </Row>
-                                    {/* Close button */}
-                                    <Row className="mt-3">
-                                        <Col className="d-flex justify-content-end">
-                                            <button type="button"
-                                                className="primaryButton fs-4 px-3"
-                                                onClick={() => {
-                                                    if (!selectedAnnotation.body.length) {
-                                                        /* Remove when annotation was left empty */
-                                                        RefreshAnnotations();
-                                                    }
-
-                                                    setSelectedAnnotation(null);
-
-                                                    anno.state.selection.clear();
-                                                }}
-                                            >
-                                                {selectedAnnotation ?
-                                                    <span> Close </span>
-                                                    : <span> Cancel </span>
-                                                }
-                                            </button>
-                                        </Col>
-                                    </Row>
-                                </Col>
-                            </Row>
-                        </div>
+                            RefreshAnnotations={() => RefreshAnnotations()}
+                            SubmitAnnotation={(values: string[], method: string) => SubmitAnnotation(values, method)}
+                            SetSelectedAnnotation={(selectedAnnotation: ImageAnnotation) => setSelectedAnnotation(selectedAnnotation)}
+                            SetEditAnnotation={(editAnnotation: ImageAnnotation) => setEditAnnotation(editAnnotation)}
+                            RemoveAnnotation={() => RemoveAnnotation()}
+                        />
                     )}
                 />
             </OpenSeadragonAnnotator>
